Tweaked Wiz build to be awesome again.
Underclocked the Wiz to 450 MHz.
Defined CRUDE_TIMING for Wiz which causes the game to run at the
exact speed it was designed for (50 fps).
Added PLATFORM_CRUDE_TIMING_TICKS which ideally should be 20.
However, for Wiz, a value of 15 works better, this could be
due to inaccuracy of SDL_GetTicks on wiz.