Fix segfault, add mock symbols.
If a brick landed on rm but a brick of same type was next to it,
it would be added to the removelist twice.
Added function to queue brick-removals which checks that a brick is not
already queued.

Tried something with the symbols for the new bricks, hopefully ViperMD
will tell me it looks like crap and make something awesome :P