Improved settings.ini and allow-internet screen.
There are now comments on all settings in settings.ini.
The wall-of-text has been reduced in size and complexity
for the "Allow online" screen which is displayed on PC on first
start, to ask the user if they allow Wizznic to go online to
check for updates, download DLCs and send helpful gameplay stats.
The uploadstats config option has been renamed to allowonline and
the showweb option is now written to the default config file.