Added crude -record function.
This recording function stores each frame in memory as TGA data
until program exit when it is saved to disk.
This uses a lot of memory and is not suitable for long recordings,
however, given that I just want this to be able to create a few
videoes showing off Wizznic, it is acceptable for me.

castawesome is the only linux screenrecorder I've had any luck with,
and it can not save to local files, so I had to build that functionality
into the game itself.