Try to fix isssue #17 (crash when moving brick)
List iteration read free'd memory after removing a block,
this seems to function on some machines and not others.

The fix is to update the iterator when deleting, and also ensure that
the current iterator is not the last item, before next iteration.

Valgrind reveals some more memory issues which I will address in later commits.